Added GitHub workflow to build and test Go code.
This CL adds a GitHub workflow [1] that builds and tests our code every
time it is pushed. I mostly copied the instructions in [2] and made
some slight modifications.

Some interesting features:

- We build and test with three go versions: 1.18, 1.19, and 1.20.
- We cache golangci-lint, so it installs very quickly.

=== Details ===

I changed allmodules from using sh to using bash. On our workstations,
sh is bash, but on the GitHub VMs, it's not. This was causing allmodules
to fail.

[1]: https://docs.github.com/en/actions
[2]: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-go