Added a skeleton blog to the website.
This PR adds a very, very minimal blog section to the website. For now,
it contains two fake blogs just to exercise the code. I took a lot of
inspiration from [Go's blog][1].

For now, there's a lot of boilerplate and repetition in writing blogs.
We could augment docgen.go to automate some of this, similar to [what
Jekyll has][2]. But, since I don't anticipate us writing a ton of blog
posts really quickly, I'll defer this until we actually need it.

[1] https://go.dev/blog/all
[2]: https://jekyllrb.com/docs/step-by-step/08-blogging/