Made website more tutorial-like.
This PR makes the website more tutorial-like, in the sense that it
should be easy for people to follow along a given section from top to
bottom.

Specifically, I revamped all OVERVIEW and FUNDAMENTALS sections.

- I mentioned that Go 1.18+ needs to be installed.
- I renamed "Getting Started" to "Step by Step Tutorial", inspired by
  https://jekyllrb.com.
- I added links to the example source code.
- I expanded code snippets to be more stand alone.
- I avoided back referencing to the "Step by Step Tutorial" section
  unless needed.
- I added a Config subsection to the Components section.
- I fixed various stale things.

In a future PR, I'll update the DEPLOYERS sections to be more standalone
as well.