Made weaver generate enforce interface subtyping.
To define a component, developers declare the interface and
implementation of the component as a go interface and struct
respectively. They also embed a weaver.Implements struct to indicate the
link between the two.

    type Adder interface {
        Add(context.Context, int, int) (int, error)
    }

    type adder struct {
        weaver.Implements[Adder]
    }

Before this PR, "weaver generate" was not enforcing that the
implementing struct (e.g., adder) actually implemented the interface
(e.g., Adder). Worse, it was not caught by "go build" either (that
surprised me). Instead, an application would crash at runtime.

This PR augments "weaver generate" to enforce that a component struct
implements the corresponding component interface.