Added --watch flag to docgen.
This PR adds a --watch flag to the website generator. When the --watch
flag is provided, the generator automatically rebuilds the website
whenever the source files are changed. This makes it much much more
convenient to the edit the website and see your changes. Before this PR,
you had to rerun the generator every time you made a change.