Added "Deploying to the Cloud" section to website.
This PR adds a "Deploying to the Cloud" section to the "Step by Step
Tutorial" on the website. The main goal of the section is to impress
upon readers that

1. you can deploy applications to the cloud easily, and
2. the deployment is really fancy and cool.

The section isn't meant to exaplain the nitty gritty details of how to
deploy the application to the cloud. That comes later in the GKE
section.

Thanks @rgrandl for the suggestion. I think this makes it much clearer
that Service Weaver can be used to run distributed applications; it's
not just for running things locally.