Updated DEPLOYERS/REFERENCE sections on website.
This PR freshens up the DEPLOYERS and REFERENCE sections on the website.

A question for reviewers. Currently, when you run `weaver gke status` or
`weaver gke-local status`, we show an estimate of the rollout schedule.
With some recent changes, these schedules are inaccurate (they've always
been inaccurate actually, but now they're really inaccurate). Should we
just remove the schedules from the status commands and the website?