Fixed "weaver generate" varargs bug.
Before this PR, "weaver generate" was not properly handling component
methods with variadic arguments. For example:

    type MyComponent interface {
        Foo(context.Context, ...int) error
    }

"weaver generate" would generate code that does not compile.
Specifically, it was treating the variadic argument `...t` as a slice
`[]t`. They're similar but not quite the same.

This PR fixes the bug. "weaver generate" now handles variadic arguments
correctly.