Attempted to fix conn_test.go flake.
The TestMetricPropagation test inside conn_test.go was flaking on GitHub
Actions (e.g., [1] and [2]). I wasn't able to reproduce the error
locally, but looking through the code, I think I might understand the
flake. This PR attempts to fix it.

TestMetricPropagation (and other similar tests) sets up an EnvelopeConn
and a WeaveletConn. These two communicate over a set of OS pipes. They
both repeatedly read a message from the pipe (blocking if needed),
handle the message, and reply on the pipe.

At the end of the test, the pipes are closed. The tests were failing
because the Conns were receiving io.EOFs. Playing around with pipes,
however, I think this is expected. If you block reading on a pipe and
the other end of pipe is closed, you get an io.EOF. The tests were
checking for io.ErrClosedPipe, but from my experimenting, it seems you
get that error only when you close *your* end of the pipe, not the other
end.

This PR checks for and ignores io.EOFs.

[1]: https://github.com/ServiceWeaver/weaver/actions/runs/4168933069
[2]: https://github.com/ServiceWeaver/weaver/actions/runs/4168814674