Added "weaver help" subcommand.
This PR adds a "weaver help" subcommand. For example, "weaver help
generate" shows a help message for the "weaver generate command". The
help command takes one of the following forms:

    weaver help
    weaver help generate
    weaver help single
    weaver help multi
    weaver help ssh
    weaver help <external>

It does not handle more complex help queries like

    // Subsubcommands aren't supported.
    weaver help multi profile

    // Flags aren't supported
    weaver help multi --foo

    // External subsubcommands aren't supported.
    weaver help gke profile

In the future, we can improve the help command to support these things
if we find it useful.