Removed examples/ dependency on gorilla mux.
This PR removes the examples/ directory's dependecy on the gorilla mux
module [1]. We replace gorilla.Router with http.ServeMux. The
replacement is mostly straightforward, but there were two small hiccups:

1. Both a gorilla.Router and an http.ServeMux allow you to dispatch
   requests based on the path of the request (e.g., "/", "/foo"). A
   gorilla.Router additionally allows you to dispatch based on the HTTP
   method (e.g., POST, GET). I implemented this logic manually (it's
   just a small helper function).
2. A gorilla.Router allows you pattern match paths and extract
   components of them. Specifically, we had a "/product/{id}" pattern
   that extracted the id of a product. Again, I implemented this logic
   manually.

This PR is part of an effort to merge the examples module with the core
weaver module without blowing up the number of dependencies.

[1]: https://github.com/gorilla/mux