Made website header fixed width across all pages.
Before this PR, the header at the top of the website, represented as a
<header> HTML element, was nested inside a <div class="page"> element
where the "page" CSS class defined the width of the page. In PR #37, we
narrowed the page width for blog posts. This lead to an inconsistent
header width.

This PR fixes the problem by hoisting the <header> element out of the
"page" div. In the future, we might want to nest it again and pick a
fixed width for all pages.

Note that there's still a weird bug where the header is slightly nudged
on the blog page. I can't figure out why it's doing that, but we should
fix it in the future too.