Handle errors
To prevent a warning from errcheck linter about unhandled errors.

The error is written to the response, and the caller doesn't use the error, so returning
an error seems to provide no benefit.

Handle the error by returning it

Print the error from terminateDeployment

Handle error from updateEndpoints

Either log or return the error. The error from updateEndpoints is not
terribly interesting at this time (only errors when the connection is
closed), but in the future it may return more interesting errors that
should be handled.

Ignore some errors

These errors don't need to be handled, so adding nolint directive to
tell the linter to ignore them.

Also fix a small bug in Fprintln using a format directive which will not
display as expected.

Log an error from receiving logs

To help debugging when logs are not delivered as expected.

Upgrade golangci-lint to latest patch version

To prevent a false positive from the nolintlint linter.

Ignore errors writting HTTP responses

Failing to write the HTTP response generally indicates that the client
is gone. Since the client is gone they don't care about the error.

The operator generally also won't care about these errors, so there is
not much value in logging them.  They should be safe to ignore.

Panic on template error

Any errors from rendering a template should be programming errors, so a
panic seems appropriate in this case. Since this is an HTTP handler the
panic will be handled by http.Server and logged to stderr.

Log the error from Unregister to stderr