Fix the copy button in the dashboard (#88)
* Fix the copy button in the dashboard

The copy button was writting an empty string to the clipboard. This
function runs when the page is first loaded, and it appears that
innerText is empty at that time, where as textContent has the appropriate
text.

* Fix text shifting position on hover

Before this change the text of the command would move by a pixel or two
on hover. By removing the padding around the copy button the text no
longer moves on hover.

---------

Co-authored-by: Robert Grandl <rgrandl@users.noreply.github.com>