Fixed some visual bugs on mobile website.
This PR fixes two bugs that appeared when the website was viewed on
mobile. First, the jumbotron on the landing page was squeezed very thin.
I fixed this by removing the left and right margins on small pages.
Second, the top navigation header had elements overlapping vertically. I
fixed this by adjusting padding.