Replaced Weavelet proto with concise WeaveletInfo. (#114)
Before this PR, we had a Weavelet proto with an embedded Deployment
proto, which in turn had an embedded AppConfig proto. When you launched
a weavelet, you passed it a Weavelet proto over the pipe. However, the
Weavelet proto contained a lot of unnecessary fields that the weavelet
never actually used. For example, does a weavelet need to know its
binary? No. This made it hard to know which fields were needed and which
weren't.

This PR replaces the Weavelet proto with a simpler WeaveletInfo proto
that only contains the fields a weavelet actually uses. With this
change, the Deployment proto becomes more of a deployer implementation
detail. We could probably remove it from the runtime package if we
wanted to, though I'll defer that.

In future CLs, I think we can make the WeaveletInfo proto even simpler.

Co-authored-by: Srdjan Petrovic <spetrovic@google.com>