Move a function to weavertest package
This is a first step in refactoring. Moving the function to the package
where it is called.

Also handle the errors from envelope Run

Fix data race

Now that ctx is read from inside the goroutine it can race with the line
below that modifies the ctx variable.

This commit fixes the problem by storing the modified context in a
different variable. This way the goroutine read of context does not
race.