Removed StartColocationGroup from the pipe API.
Consider a component A that calls weaver.Get on a different component B.
Before this PR, the weavelet hosting A would send two messages over the
pipe to the envelope.

1. It would send a StartComponent message indicating that the B
   component should be started by B's colocation group.
2. It would send a StartColocationGroup message indicating that B's
   colocation group should be started, if it hasn't already.

This PR deletes the unneeded StartColocationGroup message. When a
weavelet sends a StartComponent message, the envelope now registers that
the component should be started *and* starts its colocation group, if it
hasn't already. This simplifies the pipe API and makes implementing
deployers a little bit easier.

I also removed some unneeded fields in the StartComponent proto.