Remove the circular dependency between babysitter and Envelope
Previously both held a reference to each other.

This commit removes the collectMetrics method from babysitter and moves
it to a run method on a new metricsCollector type.

This allows the babysitter to be purely an implementation of
envelope.EnvelopHandler.

Removes 2 fields from babysitter that are no longer used after the above
change.