Fixed UseLocalhost bug introduced in PR #115.
Before PR #115, the WeaveletInfo proto had a UseLocalhost field. When
true, weavelets would listen on localhost for external listeners (the
kind returned by the Instance.Listener method). In PR #115, I removed
the UseLocalhost field and replaced it with a GetAddress API.

However, I didn't actually remove the field, and the field was being
used for more than I initially thought. The UseLocalhost field was still
in WeaveletInfo and was being used for internal listeners (the kind that
a weavelet uses to execute RPCs). The PR changed 'weaver multi' to not
set the UseLocalhost field which led to weavelets listening on $HOSTNAME
rather than localhost.

This PR fixes the bug. I replace UseLocalhost with SingleMachine since
it parallels the existing SingleProcess field better and I think it's a
bit clearer (use localhost for what?).

I also removed some related unused fields from the WeaveletInfo proto.

Fixes #154.