Wrote "How to Implement a Deployer" blog.
This PR introduces a "How to Implement a Deployer" blog which explains
what deployers are and how to write them. While writing this blog, we
discovered many things about the deployer API that could be improved
which led to a multi-week effort to revise the API, culminating in the
release of weaver v0.2.0.

Thanks to Sanjay, Srdjan, Robert, and Nipun for their great feedback on
the blog. It improved a ton thanks to everyone's input.