Only run cloud build when pushing to main.
Before this PR, a GitHub action built the code whenever it was pushed or
whenever a pull request was created. If you pushed a branch and opened a
pull request for it, all the checks would run twice. This PR changes the
GitHub action to run when pushed to main or when a pull request is
opened. This should prevent the actions from running twice.

See [1] for more information.

[1]: https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#push