Added a footer to the website for spacing. (#163)
Before this PR, our website had no padding at the bottom. This meant
that the end of page was at the very bottom of the screen. This made
reading the end of a page---like the end of a blog post---a little
irritating. This PR adds a footer to the bottom of every page. In the
future, we may want to add some contents to the footer, but for now it
exists just to add a bit of space.

Co-authored-by: Robert Grandl <rgrandl@users.noreply.github.com>