Cleaned up the multi/weavertest babysitter.
This PR cleans up the babysitter used by weaver multi and weavertest.
The changes don't impact the behavior of the babysitter; they simply
clean up the code. I made these changes in preparation for some bigger
changes I'm making in an effort to simplify the deployer API.

I made the following changes:

- I replaced the babysitter's use of versioned.Map with a new
  versioned.Versioned abstraction (first introduced in #162).
  Previously, a babysitter was storing a single piece of state in a map,
  which was a bit confusing. The new Versioned abstraction allows for
  the next change.

- I introduced a group struct to store all the information about a
  group. Previously, the state was spread across a couple different
  fields, and in some versioned.Maps.

- I removed babysitter.proto. I realized that there was no need to be
  using protos. Now, everything is vanilla Go.

- I moved the code in profile.go to babysitter.go, so the babysitter is
  now a single file.

Ironically, a lot of the changes introduced in this PR will be deleted
in future PRs as we transition to the deployer API, but these changes
will help that transition.