Implemented weavertest multiprocess deployer.
Before this PR, 'weaver multi' and weavertest shared the same
babysitter. Both 'weaver multi' and weavertest deployed applications
across multiple processes, so it made sense not to duplicate the code.

However, while we weren't duplicating code, we were also muddying the
abstractions of the babysitter, as everything had to work for both
'weaver multi' and weavertest. A good example of this is how we dealt
with pipes. For a multiprocess weavertest, the main component runs in
the same process as the babysitter. The babysitter doesn't expect this,
so weavertest created its own EnvelopeConn and used a babysitter as an
EnvelopeHandler. This worked but it wasn't easy to understand why. It
really broke the abstraction of the babysitter.

This PR implements a new weavertest multiprocess deployer. The deployer
does duplicate *some* code from the old babysitter, but much of the
babysitter's implementation wasn't needed (e.g., code for handling
metrics, routing, tracing). The new weavertest deployer has a number of
improvements:

- Cleaner handling of pipes. Now, the deployer is very much aware that
  the main component is running in the same process. I also reworked the
  Bootstrap API to avoid having files lingering around forever.
- Better teardown. The 'weaver multi' babysitter doesn't expect to get
  shut down, but a weavertest deployer does. The new deployer is more
  careful about killing all subprocesses and stopping all goroutines
  when a unit test ends.
- We're in the process of cleaning up the deployer API. As I was
  implementing some of these changes, I realized they made it very hard
  to share code between 'weaver multi' and weavertest. This was again
  because the abstractions were being tortured to work across the two.
- Enables future testing innovations. We've been toying with the idea of
  implementing more sophisticated testing. For example, we may randomly
  fail a component. Now that weavertest has its own deployer, these
  changes will be much easier to make.