Cleaned up SSH deployer's manager.
This PR cleans up the SSH deployer's manager. Specifically, I ported the
changes we made to the 'weaver multi' babysitter in #172 and #174. These
cleanups will make it easier to make some of the deployer API changes
we've been working on.

I noticed that the SSH babysitter re-uses a lot of the same code as the
'weaver multi' babysitter. I don't think we should merge the two into
one mega babysitter. In fact, I recently split the 'weaver multi'
babysitter from the weavertest babysitter, so I think there's value in
having things separate. But, maybe we can pull out some of the common
code without having a frankenstein babysitter API.