Removed co-location groups from bulk of pipe API.
Recall that, before this PR, the deployer API had four methods related
to weavelets and components: RegisterReplica, StartComponent,
GetComponentsToStart, and GetRoutingInfo. The notion of a co-location
group was baked into these APIs. A weavelet would ask for a group's
routing info, not a component's, for example.

This PR removes the idea of a co-location group from the APIs and makes
them component-centric instead. This is desirable because co-location
groups are implementation details and should not be forced into the API.
Deployers are still free to use the notion of a co-location group
internally---and right now, all our deployers do---but they do not have
to.

Note that this change does not eliminate co-location groups from the API
entirely. There are still some message that reference groups, and the
weavelet still uses the idea of a group. Future changes to the pipe API
should eliminate these entirely.