Made envelopes push components and routing info.
Before this PR, a weavelet issued blocking gets [1] to the envelope to
learn which components to run and to get updated routing information.
This version-based long polling API was complicated to explain. People
had to understand what long polling was and learn the details of our
version semantics. I also felt that it was a bit tricky to implement a
deployer using this API.

This PR replaces the long polling with envelope-initiated RPCs:
UpdateComponents and UpdateRoutingInfo. Now, whenever the set of
components changes or a component's routing info changes, the deployer
pushes an update to the weavelet.

I think this API is simpler, but a deployer implementer does have to be
careful not to forget to call UpdateComponents or UpdateRoutingInfo.
This can be tricky when events happen in unexpected orders. For example,
a weavelet may subscribe to a component before the component exists.

With this API, our weavelet also has to be careful about how it handles
UpdateComponents and UpdateRoutingInfo. In general, a weavelet handler
should never block or touch the pipe. Doing so makes it easy for a
deployer to deadlock. For example, imagine an envelope receives a
StartComponent request. It grabs a lock and then issues an
UpdateComponents request. If the weavelet then issues another RPC over
the pipe, and the envelope handles this RPC by grabbing the same lock,
there is deadlock.

=== Future Work ===

Note that there are still lingering artifacts in the pipe API. There
are still co-location groups here and there despite the work in PR #181.
The weavelet still assumes placement based on groups. I'll continue to
fix these issues in future PRs.

=== Details ===

I removed versioned.Versioned from the weavertest and weaver multi
deployers. The SSH deployer still uses it, but we can remove that in the
future.

[1]: https://javascript.info/long-polling