Allow Envelope to be used directly in testing
Previously EnvelopeConn was used by some tests and weavertest. This
commit allows Envelope itself to be used by those tests by smuggling the
reader and writer through the context.

This allows ReadMetrics and metrics.Importer to move to Envelope.

This change is being make so that EnvelopeConn can be unexported and
removed from a public interface.