Unflaked weavelet race condition.
Before this PR, there was a race condition in starting the weavelet. The
weavelet created a WeaveletConn with itself as a handler. This
WeaveletConn would immediately start serving requests and calling into
the handler. However, the handler (i.e. the weavlet) wasn't yet fully
constructed. There was a race between initializing some fields of the
weavelet and those fields being used by the handler.

This PR fixes the race by delaying when the WeaveletConn starts serving
until after the weavelet is fully initialized.