Added GitHub action for uncommitted changes.
'./dev/build_and_test generate' generates and modifies a number of
files. It generates code for all the protos, it generates
'weaver_gen.go' files, it updates godeps.txt, and it adds licenses to
the tops of files. You're supposed to run './build_and_test generate'
before opening a PR, but it's easy to forget (I forget all the time).

This PR adds a GitHub action to check that you've run './build_and_test
generate'. Specifically, it runs the script and makes sure that no file
has changed.