Replace the unsafe use of the runtime.FuncForPC
Based on the slog, it's not safe to use rec.PC as an argument for
runtime.FuncForPC.

We use runtime.FuncForPC to retrieve the filename and the line number
for logging.

As dnephin@ suggested, there is another way to do the same thing, w/o
using runtime.FuncForPC.

Thanks dnephin@ for pointing out this issue.