Added prototype versioning system.
> Version Issues

Service Weaver is currently at v0, which means we're making a lot of
breaking changes to the code. This has the problem that a deployer and
the app its trying to deploy may be at incompatible versions.

For example, imagine someone is trying to learn Service Weaver. They run
`go install github.com/ServiceWeaver/weaver@latest`. A couple weeks
later, they start writing a new Service Weaver application and run `go
mod tidy`. `go mod tidy` picks up the latest version of Service Weaver
which is likely incompatible with the stale version of the `weaver` tool
they downloaded a couple weeks back. As a result, the `weaver` tool will
likely fail to deploy the application, and it won't produce any
meaningful error messages explaining why it failed.

> Proposed Solution

This PR introduces a proposed solution. It's not perfect, but hopefully
it kickstarts some ideas on how to solve this annoying versioning issue.
With this PR, we version the deployer API. A weavelet reports to the
envelope the deployer API version it was built with. The envelope can
then check whether it supports the reported version, aborting if not.
See runtime/version.go for more details.