Added "weaver version" command.
> Example

```console
$ weaver version
weaver version: commit=254d063fcc5a3c27f95869215dfa6bb36bbfbac6 deployer=v0.2.0 target=linux/amd64
```

> Overview

This PR introduces a `weaver version` command which shows the version of
the `weaver` tool. The output includes the git commit at which the tool
was built, the deployer API version, and the target OS and architecture.

> Motivation

As described in #219, a Service Weaver deployer and the app it is trying
to deploy need to have compatible versions. We're seeing some issues
(e.g., #179, #154) that I think may be cause by version issues. `weaver
version` makes it easier to debug these kinds of issues, as people can
include the output of `weaver version` in their bug reports.