Move listener close to server
This matches the behaviour of http.Server, and helps keep the tested
behaviour in line with the real production behaviour.

Previously TestCancelServe was only working because the listener was
cancelled in the test. Without that call to cancel the context being
cancelled did not actually stop Serve.

Also remove duplicate errors.Is condition and fix formatting of message
formats comment.