Use balancer from ClientOptions
Originally I wanted to move the initialization of opts.Balancer into
withDefaults, however that caused some tests to fail because the client
had a second reference to the balancer.

This commit removes the extra reference. The lock continues to guard
calls to balancer.Pick, but the balancer reference itself is never
modified after the type is initialized.