Remove call.NetworkAddress
This type was only used as a way to parse a call.NetEndpoint. We can
replace the type with a ParseNetEndpoint function.

Also use field names for constructing a slice, so that it's easier to
find where replicaSet is initialized. Without these field names an IDE
may not be able to find this initialization. I've run into this problem
twice now.