Replaced ErrRetriable with RemoteCallError.
> Retrying, The API

Before this PR, we had an error called `weaver.ErrRetriable`. The idea
was that component method calls would sometimes return an error with an
embedded `ErrRetriable`. You could detect this error with `errors.Is`
and retry the failed method call:

```go
var val string
var err error
for i := 0; i < 5; i++ {
    val, err = cache.Get(ctx, "key")
    if errors.Is(err, weaver.ErrRetriable) {
        // Retriable system error! Retry.
        continue
    }
    break
}
```

In theory, this is great, but what exactly does it mean for an error to
be retriable? Does it mean that the method call was guaranteed to never
have started executing? If so, why don't we just retry the error for the
user? Or does it mean that the method call was maybe partially or fully
executed? But then, it might not be safe to retry the call. And what
about system errors that are not retriable? How do I detect those?

> Retrying the API

This PR throws out `ErrRetriable` and replaces it with a much simpler
and more straightforward `weaver.RemoteCallError` error. A
`RemoteCallError` indicates that a component method call failed to
execute properly (e.g., because of a failed machine or a network
partition). That's it. It doesn't imply anything about the retry-ability
of a call. Here's an illustrative example:

```go
// Call the foo.Foo method.
err := foo.Foo(ctx)
if errors.Is(err, weaver.RemoteCallError) {
    // foo.Foo did not execute properly.
} else if err != nil {
    // foo.Foo executed properly, but returned an error.
} else {
    // foo.Foo executed properly and did not return an error.
}
```

It is up to the user to retry failed calls carefully. Actually, this was
always true, but I think the new API makes that clearer. Internally, if
we detect that a method call failed and we are 100% sure it never
started executing (e.g., because we failed to make a network connection
in the first place), we can retry the cal transparently for the user.

The new API is also easier to implement and allows us to remove some
ugly code that I had to introduce to avoid some cyclic dependencies.

In the future, we can extend the API with a set of canonical error codes
if we want, similar to [1].

[1]: https://cloud.google.com/apis/design/errors