Added reverser example app, derived from wrap.
In PRs #235, #236, and #237, I added an example wrap app that wrapped
text. Our plan was to use this app in a demo, explaining how to write
and deploy applications. Additionally, we would host an instance of the
app on GKE and let people play with it.

After implementing the app, we realized that it was a bit confusing. The
text you wrote was getting wrapped naturally by the browser. The text
returned by the Wrapper component was also getting wrapped, but in a
different way that was imperceptible unless you understood what was
going on. Plus, the implementation got a little too complicated.

This PR introduces a much simpler reverser app. It borrows the same UI
from the wrap app but reverses text instead of wrapping it, which is
easier to understand. The implementation of reversing is also much
simpler than wrapping.

We'll continue to iterate on the example app. If we like this one, we
can delete the wrap app in the future.