Added a blog post on Service Weaver vs. CORBA.
We've been hearing a lot of feedback along these lines:

> Service Weaver hides the complexities of remote method calls. This
> violates the fallacies of distributed computing and is destined to
> fail. CORBA failed for this exact reason. Aren't you doomed to repeat
> the past?

There's a lot to unpack in such feedback. A lot of it is true, and some
is not. We decided to write a blog post to clearly explain how Service
Weaver deals with the fallacies of distributed computing and how it
relates to CORBA.