Cleaned up weaver generate's static analysis code.
> `weaver generate` Refresher

`weaver generate` has two phases. In the first phase, it analyzes and
validates the provided code. For example:

- It finds every component interface and implementation.
- It checks that the implementation implements the interface and that
  all method arguments and returns are serializable.
- It finds every router type and checks that the router's methods match
  the component's methods.

In the second phase, it generates code and places it in weaver_gen.go
files.

> Changes

Before this PR, the `generator` type was responsible for both validation
and generation. This PR changes validation to use pure functions and
makes `generator` responsible solely for generation. I also made a bunch
of miscellaneous cleanups.

Now that validation is performed as pure functions, the flow of errors
is more explicit. Before, all errors were placed in `generator.errors`,
which made it a bit confusing to what was failing. For example, code
generation almost never fails. I don't think this was clear before.

A question to reviewers, would you prefer we place all the validation
code in a separate file? Right now, generator.go is pretty long (2100+
lines).