Allow empty component interfaces.
This PR changes `weaver generate` to allow empty component interfaces.
We're in the process of converting main to be an explicit component that
implements an empty `weaver.Main` interface. This change will allow
that. We've also received [an issue][issue] explaining a use case for an
interface without any methods.

[issue]: https://github.com/ServiceWeaver/weaver/issues/270