Fixed annoying website wiggles.
Before this PR, the website header would sometimes shift slightly left
or right depending on which page you were on. This caused the site to
wiggle if you switched back and forth between two pages.

I spent some time debugging this. At first, I thought it was some super
complicated CSS subtlety related flexbox or something. After a while, it
dawned on me that it was much simpler. On pages that were long, the
scrollbar appeared, which slightly shrunk the width of the page, causing
the header to move left.

This PR fixes the problem by setting the body width to 100vw.